% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{addl}
\alias{addl}
\title{Create ADDL DATA}
\usage{
addl(
  data = dat,
  id = "id",
  time = "time",
  last.dose.time = "last.dose.time",
  ii = 24,
  note = "use only dose data with evid=1."
)
}
\arguments{
\item{data}{Data frame.}

\item{id}{subject id}

\item{time}{Time after first dose}

\item{last.dose.time}{This could be minimum last PK time. Required to compute the ADDL of the last dose}

\item{ii}{Dose interval vector}
}
\description{
Create ADDL DATA
}
\keyword{addl}
