% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{AUC}
\alias{AUC}
\title{Calculate AUC Using the Trapezoidal Method}
\usage{
AUC(data = dat, time = "TIME", id = "ID", dv = "DV", method = "linlog")
}
\arguments{
\item{time}{chronologically ordered time variable present in data}

\item{id}{variable in data defining subject level data}

\item{dv}{dependent variable used to calculate AUC present in data}

\item{method}{linlog= lin up and log down}

\item{data.frame}{containing the data to use for the AUC calculation}
}
\description{
data
}
\examples{
AUC(data, time = 'TIME', id = 'ID', dv = 'DV')
}
\keyword{AUC}
