% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{bsv_add}
\alias{bsv_add}
\title{Derive BSV and RSE with additive ETA}
\usage{
bsv_add(theta = 4.25546, omega = 0.495652, se_omega = NULL, fancy = "yes")
}
\arguments{
\item{omega}{omega}

\item{se_omega}{se omega}
}
\description{
Derive BSV and RSE with additive ETA
}
\keyword{bsv_add}
