% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{ci_bin}
\alias{ci_bin}
\title{Derive confidence interval for binary data}
\usage{
ci_bin(p, n, ci = 0.975)
}
\arguments{
\item{p}{probabililty}

\item{n}{total number}

\item{ci}{confidence interval 0.975 for 95%. To derive lower and upper bounds, do p -/+ ci_bin}
}
\description{
Derive confidence interval for binary data
}
\keyword{ci_bin}
