% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{ci_prob}
\alias{ci_prob}
\title{Derive confidence interval for binary data}
\usage{
ci_prob(n, total)
}
\arguments{
\item{n}{number of responder}

\item{total}{total number}
}
\description{
Derive confidence interval for binary data
}
\keyword{ci_prob}
