% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{comp_fun}
\alias{comp_fun}
\title{Compare Function}
\usage{
comp_fun(x, y, type = "var")
}
\arguments{
\item{x}{data 1}

\item{y}{data 2}

\item{type}{If var for variable else data frame}
}
\description{
Compare Function
}
\keyword{comp_fun}
