% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{dummy.fun}
\alias{dummy.fun}
\title{Dummy functions for code template}
\usage{
dummy.fun(dummy)
}
\arguments{
\item{dummy}{Data frame. Note, it's not a working function but codes could be used. See tips.plot in lhplot as well for other templates.}
}
\description{
Dummy functions for code template
}
\keyword{dummy.fun}
