% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{errprop}
\alias{errprop}
\title{Compute Error Propagation x and y}
\usage{
errprop(
  x = c(0.5, 0.1),
  y = c(0.5, 0.1),
  exp = expression(sqrt(exp(x) - 1) + y),
  sim = F,
  verbiose = F,
  nsim = 1e+05,
  raw = T
)
}
\arguments{
\item{x}{Pair value and sd for x}

\item{y}{Pair value and sd for y. If y not used, set it to value and sd to 0}

\item{exp}{expression(equation)}

\item{sim}{True if simulation is used to derive the error.}
}
\description{
Compute Error Propagation x and y
}
\keyword{errprop}
