% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{hl1cpt}
\alias{hl1cpt}
\title{One compartment micro constants and HL}
\usage{
hl1cpt(cl, v)
}
\arguments{
\item{data}{data}
}
\description{
This function allows you to derive TAD and RTIME from calendar date/time.
}
\keyword{hl1cpt}
