% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{hl2cpt}
\alias{hl2cpt}
\title{Two compartment micro constants and HL}
\usage{
hl2cpt(cl, cl2, v, v2)
}
\arguments{
\item{data}{data}
}
\description{
This function allows you to derive micro constants and HL.
}
\keyword{hl2cpt}
