% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh.def}
\alias{lh.def}
\title{Create define or data specification}
\usage{
lh.def(lab = c("code;;define;;unit", "b;;test>=b;;ug", "c;;test<c;b;a;; "))
}
\arguments{
\item{lab}{data specification c(variable name;;label;;unit)}
}
\description{
Verify if vector is varying or duplicate
}
\keyword{lh.def}
