% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh.ehl.rc}
\alias{lh.ehl.rc}
\title{Derive Effective Half-life and Accumulation Ratio}
\usage{
lh.ehl.rc(data, AUCsd = "AUCsd", AUCss = "AUCss", TAU = 24)
}
\arguments{
\item{data}{NCA results}

\item{AUCsd}{AUCtau SD}

\item{AUCss}{AUCtau SS}

\item{TAU}{Dosing iterval}

\item{id}{unique subject identifier}

\item{OCC}{Identifier of SD and SS}
}
\description{
Require AUC of first dose, AUCtau and tau
}
\examples{
t<-lhwide(test[,c("id","OCC","AUClast")],"AUClast","OCC")
names(t)<-c("id","AUCsd","AUCss")
lh.ehl.rc(data=t,AUCsd="AUCsd",AUCss="AUCss",TAU=24)
}
\keyword{lh.ehl.rc()}
