% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh.join}
\alias{lh.join}
\title{lhjoin funtion
Join two datasets and print report of joining procedure}
\usage{
lh.join(dat1, by1 = NULL, dat2, by2 = NULL, type = "full")
}
\arguments{
\item{dat1}{by1 Data frame 1 and variables to be matched. If NULL, match="all"}

\item{dat2}{by2 Data frame 2 and variables to be matched. If by1=NULL then by2=NULL then match="all"}

\item{type}{could be "full", "left","right" or "inner"}
}
\description{
lhjoin funtion
Join two datasets and print report of joining procedure
}
\keyword{lhjoin}
