% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_data_spec}
\alias{lh_data_spec}
\title{Create define or data specification}
\usage{
lh_data_spec(
  lab = c("code<>define<>unit", "b<>test>=b<>ug", "c<>test<c;b;a<> ")
)
}
\arguments{
\item{lab}{data specification c(variable name;;label;;unit)}
}
\description{
Verify if vector is varying or duplicate
}
\examples{
lh_data_spec(lab)
}
\keyword{lh_data_spec}
