% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_duration}
\alias{lh_duration}
\title{Compute duration below and above target}
\usage{
lh_duration(
  data = simd,
  time = "TIME",
  dv = "CONC",
  y.cutoff = 80,
  id = "ID",
  caryover = c("DOSEN")
)
}
\arguments{
\item{data}{input data}

\item{time}{time}

\item{dv}{dependent variable}

\item{y.cutoff}{cutoff value}

\item{id}{unique subject identifier}

\item{caryover}{variable in data to be carried over with the output}
}
\description{
Compute duration below and above target
}
\keyword{lh_duration}
