% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_expand}
\alias{lh_expand}
\title{Expand data frame}
\usage{
lh_expand(data, n.rows = 20)
}
\arguments{
\item{data}{data frame to be expand}

\item{n.rows}{number of row to be expanded}
}
\description{
Expand data frame
}
\keyword{lh_expand}
