% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_proximity}
\alias{lh_proximity}
\title{Find the closest df2 value (or date) to df1 value}
\usage{
lh_proximity(df1 = df1, df2 = df2, direction = "<")
}
\arguments{
\item{df1}{fixed vector}

\item{df2}{relative vector}

\item{direction}{keep df2 value equal to or smaller than df1 value (<)}

\item{diff}{output differences between df1 and the closest df2}
}
\description{
Find the closest df2 value (or date) to df1 value
}
\keyword{lh_proximity}
