% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_rmd_tem}
\alias{lh_rmd_tem}
\title{Dummy functions for Rmarkdown}
\usage{
lh_rmd_tem(dummy = "dummy")
}
\arguments{
\item{dummy}{Data frame. Note, it's not a working function but codes could be used.}
}
\description{
Dummy functions for Rmarkdown
}
\examples{
  ---
  title: "Exposure-Response Modeling and Simulation of Apitegromab in Spinal Muscular Atrophy Patients (RRK-PMX-APITEGROMAB-3225)"
  author: "Leon Pheng"
  date: "`r format(Sys.time(), '\%a \%b \%d \%Y \%X')`"
  output:
    html_document:
    code_folding: hide
  df_print: paged
  fig_caption: yes
  fig_height: 12
  fig_width: 12
  number_sections: yes
  theme: united
  toc: yes
  toc_float: yes
  self_contained: TRUE
  word_document:
    toc: yes
  editor_options:
    chunk_output_type: console
  ---
    ```{r setup, include=FALSE}
  knitr::opts_chunk$set(echo=T, results='asis', warning=FALSE, message=FALSE, fig.width=5, fig.height=5, fig.path="Figures/")
  ```
  ```{r, echo=FALSE}
  library(plotly);library(ggplot2);library(table1);library(tidyverse);library(readxl)
  library(knitr);library(plyr);library(lhtool2)
  ```
  # Add R command
  ```{r, echo=FALSE,include=FALSE}
  ```
  # Table data frame
  ```{r, echo=FALSE}
  d<-read.csv("./typical.csv")
  kable(d)
  ```
  # Display tabulation {.tabset}
  # Inserting png
  ![](./output/GOF_all.png)
  # Insert pdf
  ![](equations.pdf){width=100\% height=400}
  # Insert another html doc
  ```{r, echo=FALSE}
  htmltools::includeHTML("./demo.html")
  ```
}
\keyword{lh_rmd_tem}
