% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_stat}
\alias{lh_stat}
\title{Descriptive statistics using addvar function}
\usage{
lh_stat(dat, by, var, stat.fun = "mean")
}
\arguments{
\item{dat}{data frame}

\item{by}{sort list}

\item{var}{variables list}

\item{stat.fun}{basic statistic function (single variate) ex:mean, sd, se, cv}
}
\description{
Descriptive statistics using addvar function
}
\keyword{lh_stat}
