% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_template}
\alias{lh_template}
\title{Open template from source}
\usage{
lh_template(
  script = "rmd.rmd",
  path = "C:/Users/lpheng/OneDrive - Certara/Desktop/packdev"
)
}
\arguments{
\item{script}{script name}

\item{path}{source path}
}
\description{
Open template from source
}
\keyword{lh_template}
