% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_tips}
\alias{lh_tips}
\title{Dummy functions for code template}
\usage{
lh_tips(dummy = "dummy")
}
\arguments{
\item{dummy}{Data frame. Note, it's not a working function but codes could be used.}
}
\description{
Dummy functions for code template
}
\keyword{lh_tips}
