% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhcbind}
\alias{lhcbind}
\title{lhcbind}
\usage{
lhcbind(dat1, dat2)
}
\arguments{
\item{dat1}{data frames 1.}

\item{dat2}{data frames 2.}
}
\description{
C bind 2 data frames regardless number of row length.
}
\keyword{lhcbind}
