% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhflex}
\alias{lhflex}
\title{Make Flexible table}
\usage{
lhflex(
  table1,
  csv = "yes",
  bord = "yes",
  select = NULL,
  add.h = NULL,
  merge.all = "yes",
  size = 12,
  empty = NULL,
  cf = NULL,
  border = NULL,
  align = "center"
)
}
\arguments{
\item{table1}{data frame}
}
\description{
Generate descriptive statistic of continuous variable with style
}
\examples{
tab1<-lhtab1(data=dat1,sort.by="ARM",cont=continous,cat=categorical,render="word",overall="yes")
print(tab1,"Demog.docx")
}
\keyword{lhtab1(data=df,sort.by=c("study","form"),cont=cont,cat=NULL,stats="stat1",fun="fun1",overall="yes",render="flex",transpose=F)}
