% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhrbind}
\alias{lhrbind}
\title{lhrbind}
\usage{
lhrbind(dat1, dat2, na.replace = NA, all.character = T)
}
\arguments{
\item{dat1}{data frames 1.}

\item{dat2}{data frames 2.}
}
\description{
r bind 2 data frames regardless number of columns.
}
\keyword{lhrbind}
