% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtab2}
\alias{lhtab2}
\title{Descriptive Statistics Continuous with Style}
\usage{
lhtab2(
  data,
  sort.by = c("STUDYID", "SEXC"),
  cont = c("ALT", "BAST", "AST"),
  stats = c("length(x[!is.na(x)])=N", "length(x[is.na(x)])=Nmiss", "geom(x)=GeoMean",
    "median(x,na.rm=T)=Median", "quantile(x,0.5,na.rm=T)=50thPI", "mean(x,na.rm=T)=Mean",
    "cv(x)=CV\%", "min(x)=Min", "max(x)=Max", "geocv(x)=GeoCV\%"),
  stat.group = list(c("N", " (", "Nmiss", ")"), c("Mean", " (", "CV\%", ")"),
    c("Median", " [", "Min", ", ", "Max", "]"), c("GeoMean", " (", "GeoCV\%", ")")),
  render = "flextable",
  overall = "yes",
  format = "stacked"
)
}
\arguments{
\item{data}{dataset}

\item{sort.by}{sorting variables}

\item{cont}{list of continuous variables}

\item{stats}{statistic functions. stat1 contained most of basic statistic function (mean, median up to CI95). User can define personal list of function (ex: c("length(x)/mean(x)=RT"))}

\item{render}{the output format as flexible table "flex", as "csv". Note that officer package is required for the word format and save the output as docx (ex: print(doc,"table1.docx))}

\item{overall}{if the overall stats required then overall="yes"}

\item{format}{two formats available, stacked or not}

\item{cat}{list of categorical variables}

\item{fun}{define the output}
}
\description{
Generate descriptive statistic of continuous variable with style
}
\examples{
tab1<-lhtab1(data=dat1,sort.by="ARM",cont=continous,cat=categorical,render="word",overall="yes")
print(tab1,"Demog.docx")
}
\keyword{lhtab1(data=df,sort.by=c("study","form"),cont=cont,cat=NULL,stats="stat1",fun="fun1",overall="yes",render="flex",transpose=F)}
