% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtab3}
\alias{lhtab3}
\title{Descriptive Statistics Continuous in matrix style}
\usage{
lhtab3(
  data,
  sort.header = c("group", "studyid"),
  sort.body = NULL,
  cont = c("bestpcttl", "WT"),
  cat = NULL,
  render = "demog.docx",
  overall = T,
  stacked = T,
  rounding = "sigfig(x,3)",
  stat1 = c("length(x)=N", "length(x[is.na(x)])=Nmiss", "mean(x,na.rm=T)=Mean",
    "median(x,na.rm=T)=Median", "min(x,na.rm=T)=Min", "max(x,na.rm=T)=Max", "cv(x)=CV",
    "geom(x)=GeoMean", "geocv(x)=GeoCV"),
  stat.label = list(c("N"), c("Mean", " (", "CV", ")"), c("Median", " [", "Min", "-",
    "Max", "]"), c("GeoMean", " (", "GeoCV", ")"))
)
}
\arguments{
\item{data}{dataset}

\item{sort.header}{horizontal sorting}

\item{sort.body}{vertical sorting}

\item{cont}{list of continuous variables}

\item{render}{if docx or csv then render as flextable}

\item{overall}{output overall}

\item{stacked}{output the stats in stacking style for continuous}

\item{rounding}{define rounding method for continuous}

\item{stat1}{define statistic functions for continous}

\item{stat.label}{define group of statistics to be outputted}
}
\description{
Generate descriptive statistic of continuous variable with style
}
\keyword{lhtab3()}
