% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nca.cal}
\alias{nca.cal}
\title{Derive Common NCA parameters using single and multiple profiles}
\usage{
nca.cal(
  data = data,
  n_lambda = 3,
  id = "id",
  time = "time",
  dv = "dv",
  dose = "amt",
  multiple.dose = "flag",
  partialAUC = c(0, 20, 2, 14),
  partialConc = c(4, 5)
)
}
\arguments{
\item{data}{datset or data frame (ex:data=PKdatat)}

\item{n_lambda}{number of points for estimating the Lambda}

\item{id}{unique subject identifier}

\item{time}{Sampling time after dose (TAD)}

\item{dv}{Concentration}

\item{dose}{Dose administered as vector. Required to estimate CL and Vss}

\item{partialAUC}{Time interval for partial AUC. Ex: c(0,6,0,12,6,12) for AUC0-6, AUC0-12 and AUC6-12}

\item{partialConc}{Point estimated concentration (Ex:c(1,4) for concentration after 1 and 4 h)}

\item{dosing.regimen}{PK profile profile flag (ex: single, sd, md ,ss).As vector or numeric. Note that only one prefile per dose is to be analyzed.}
}
\description{
nca.cal()
}
\examples{
data<-data.frame(id=1,time=c(0,0.1,0.5,1,4,6,10,15,20,24,c(0,0.1,0.5,1,4,6,10,15
,20,24)+48),dv=c(0,0.1,0.5,10,9,8,3,1,0.5,0.1,c(2,3,4,12,10,9,4,2,0.8,0.5)),
amt=20)
data$flag<-ifelse(data$time<=24,1,2)
test<-nca.cal(data=data,n_lambda = 3, id = "id", time = "time", dv = "dv",dose
="amt",multiple.dose="flag",partialAUC=c(0,20,2,14),partialConc=c(4,5))
}
\keyword{nca.cal}
