% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nm_ph}
\alias{nm_ph}
\title{Format NONMEM outputs for phx_typical}
\usage{
nm_ph(ext = "BLOC6.ext", tab = "mod_3001_noJAP.TAB", id = "id")
}
\arguments{
\item{ext}{theta sheet}

\item{phi}{required for shrinkage calculation. Outputs: 1 for theta; 2 for omega and shrinkage; 3 for se theta; 4 for se omega}
}
\description{
Format NONMEM outputs for phx_typical
}
\keyword{nm_ph}
