% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nmid}
\alias{nmid}
\title{NMID}
\usage{
nmid(data, id = "USUBJID", varname = "NMID")
}
\arguments{
\item{data}{data frame}

\item{id}{subject id}

\item{varname}{column name}
}
\description{
This function allows you to create NMID.
}
\keyword{nmid}
