% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{phx_bsv}
\alias{phx_bsv}
\title{Compute BSV from Phoenix Omega}
\usage{
phx_bsv(par = "nCl", om)
}
\arguments{
\item{par}{Parameter names}

\item{om}{phoenix omega output, matrix}
}
\description{
Compute BSV from Phoenix Omega
}
\keyword{phx_bsv}
