% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{phx_omrse}
\alias{phx_omrse}
\title{Compute RSE for Omega from Phoenix}
\usage{
phx_omrse(
  par = "nCl",
  om,
  sdom,
  exp = expression(sqrt(exp(x) - 1) + y),
  y = c(0, 0)
)
}
\arguments{
\item{par}{Parameter names}

\item{om}{Phoenix omega output, matrix}

\item{sdom}{Phoenix omega SD output, matrix}
}
\description{
Compute RSE for Omega from Phoenix
}
\keyword{phx_shrk}
