% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{phx_shrk}
\alias{phx_shrk}
\title{Compute BSV from Phoenix shrinkage}
\usage{
phx_shrk(par = "nCl", om)
}
\arguments{
\item{par}{Parameter names}

\item{om}{phoenix omega output, matrix}
}
\description{
Compute BSV from Phoenix shrinkage
}
\keyword{phx_shrk}
