% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{phx_typical}
\alias{phx_typical}
\title{Summary of NONMEM typical values using outputs from nm_ph function}
\usage{
phx_typical(
  theta = "nm_ph1",
  omega = "nm_ph2",
  sd = "nm_ph3",
  omega_sd = "nm_ph4",
  estimate = "value",
  lab = c("tvKa;;Ka (1/h);;x+y;;c(0,0);;nKa;;sqrt(exp(x)-1)*100",
    "dKadSTR100;;If dose=100;;exp(x)+y;;c(0,0);;",
    "tvCl;;CL/F (L/h);;x+y;;c(0,0);;nCl;;sqrt(exp(x)-1)*100",
    "dCldMULT2;;CL, if multiple dose;;exp(x)*y;;tvCl;;",
    "tvV;;Vc/F (L);;x+y;;c(0,0);;nV;;sqrt(exp(x)-1)*100",
    "tvCl2;;Q/F (L/h);;x+y;;c(0,0);;nCl2;;sqrt(exp(x)-1)*100",
    "tvV2;;Vp/F (L/h);;x+y;;c(0,0);;nV2;;sqrt(exp(x)-1)*100",
    "stdev0;;Proportional Error (\%);;x*100+y;;c(0,0);;")
)
}
\arguments{
\item{theta}{theta sheet nm_ph 1}

\item{omega}{omega data matrix format nm_ph 2}

\item{sd}{set to null if no covariance steps or specify variable name for SD theta.}

\item{omega_sd}{sd for omega matrix format nm_ph 4. Null if no covariance steps}

\item{estimate}{specify variable name for estimate, default is value}

\item{lab}{table format}
}
\description{
Summary of NONMEM typical values using outputs from nm_ph function
}
\keyword{phx_typical}
