% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{ppt.function}
\alias{ppt.function}
\title{Create Power point}
\usage{
ppt.function(...)
}
\arguments{
\item{template}{Required typical template by Leon}

\item{cover}{Cover slide c("Title","Authors","Project No or Date").}

\item{plain1}{Slide with single item}

\item{plain2}{Slide with two items}
}
\description{
Create Power point
}
\keyword{pptdoc}
