% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{reflag}
\alias{reflag}
\title{Reflag variables}
\usage{
reflag(
  dat,
  var,
  orignal.flag,
  new.flag = NULL,
  newvar = NULL,
  to.factor = T,
  missing = c("", ".", "NA", NA)
)
}
\arguments{
\item{dat}{data frame}

\item{var}{Vector to be changed}

\item{orignal.flag}{Original names ex:"M","F"}

\item{new.flag}{New names ex:c"Male","Female"}

\item{newvar}{Create new vector}
}
\description{
This function allows you to change variable name ex: "M" to "Male".
}
\keyword{reflag}
