% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{tadRT}
\alias{tadRT}
\title{Derive TAD and RTIME}
\usage{
tadRT(
  data,
  id = "ID",
  date = "DATE",
  time = "CTIME",
  EVID = "EVID",
  tz = "UTC",
  format = "\%Y-\%m-\%d \%H:\%M"
)
}
\arguments{
\item{data}{data}

\item{id}{subject id}

\item{date}{date variable "%Y-%m-%d"}

\item{time}{time variable  "%H:%M")}

\item{EVID}{evid variable (evid>0 for dose)}
}
\description{
This function allows you to derive TAD and RTIME from calendar date/time.
}
\keyword{tadRT}
