% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{uc}
\alias{uc}
\title{Unicode}
\usage{
uc(symbol = "gte")
}
\arguments{
\item{symbol}{code name: ex: alpha or alfa or al, gte=greater than or egale; lte=less than or equal}
}
\description{
Unicode
}
\keyword{uc}
