% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{wdoc}
\alias{wdoc}
\title{Create word doc}
\usage{
wdoc(template = NULL, start = list(NULL))
}
\arguments{
\item{template}{Document template if available, provide full path and template file name}

\item{df}{Data frame will be converted by flextable table}

\item{tab}{Table created by flextable}

\item{txt}{Text or paragraph. For heading 1, 2 or 3 text c("text",1, 2 or 3)}

\item{img}{Plots saved as png. Enter the full path and file name}

\item{fig}{Plot created by ggplot}
}
\description{
Create word doc
}
\keyword{wdoc}
