% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lh_rmd_tem}
\alias{lh_rmd_tem}
\title{Dummy functions for Rmarkdown}
\usage{
lh_rmd_tem(dummy = "dummy")
}
\arguments{
\item{dummy}{Data frame. Note, it's not a working function but codes could be used.}
}
\description{
Dummy functions for Rmarkdown
}
\keyword{lh_rmd_tem}
