% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtab1}
\alias{lhtab1}
\title{Descriptive Statistics Continuous and Discrete}
\usage{
lhtab1(
  data,
  sort.by = c("study", "form"),
  cont = NULL,
  cat = c("Sex", "Race"),
  stats = "stat1",
  fun = "fun1",
  overall = "yes",
  render = "flex",
  transpose = F
)
}
\arguments{
\item{data}{dataset}

\item{sort.by}{sorting variables}

\item{cont}{list of continuous variables}

\item{cat}{list of categorical variables}

\item{stats}{statistic functions. stat1 contained most of basic statistic function}

\item{fun}{define the output}

\item{overall}{if the overall stats required then overall="yes"}

\item{render}{the output format as flexible table "flex", as "csv", or word document ("word"). Note that officer package is required for the word format and save the output as docx (ex: print(doc,"table1.docx))}

\item{transpose}{when transpose is TRUE, the output will be in docx containing both continuous and categorical covariate}
}
\description{
Generate descriptive statistic of continuous and/or categorical variables
}
\keyword{lhtab1(data=df,sort.by=c("study","form"),cont=cont,cat=NULL,stats="stat1",fun="fun1",overall="yes",render="flex",transpose=F)}
