% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtab3}
\alias{lhtab3}
\title{Individual table with descriptive statse}
\usage{
lhtab3(
  data = d,
  id = "animal",
  by = c("antibody_name", "dose_mpk"),
  horiz = "time_point",
  var.list = "conc",
  round = "sifig",
  dec = 3,
  heading = 1,
  stat.eq = NULL,
  stat.list = NULL
)
}
\arguments{
\item{data}{datset or data frame (ex:data=PKdatat)}

\item{id}{unique identifier}

\item{by}{Stratification variable (ex: by="study")}

\item{dec}{round decimal or number of significant figures}

\item{variables}{Specify sorting variable to be displayed vertically. (ex: colby=by or colby="var")}

\item{rtype}{rounding type. (sigfig by default)}
}
\description{
Listing of individual data and descriptive stats
}
\examples{
lhtab3(data=dat,id="NMID",by=c("study"))
}
\keyword{ind.tab}
