% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtime_var}
\alias{lhtime_var}
\title{CHECK TIME VARYING OR DUPLICATE}
\usage{
lhtime_var(data, by = "ID", var = c("BBILI", "BILI"))
}
\arguments{
\item{data}{data frame}

\item{by}{Fixed or sorted vector (s)}

\item{var}{Vector (s) to be verified}
}
\description{
Verify if vector is varying or duplicate
}
\keyword{lhtab1(data=df,sort.by=c("study","form"),cont=cont,cat=NULL,stats="stat1",fun="fun1",overall="yes",render="flex",transpose=F)}
