% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nca.cal}
\alias{nca.cal}
\title{Derive Common NCA parameters using single and multiple profiles}
\usage{
nca.cal(
  data,
  n_lambda = 3,
  id = "id",
  time = "time",
  dv = "dv",
  partialAUC = NULL,
  partialConc = NULL,
  full = F
)
}
\arguments{
\item{data}{datset or data frame (ex:data=PKdatat)}

\item{n_lambda}{number of points for estimating the Lambda}

\item{id}{unique subject identifier}

\item{time}{Sampling time}

\item{dv}{Concentration}

\item{partialAUC}{Time interval for partial AUC. Ex: 0,6,0,12,6,12 for AUC0-6, AUC0-12 and AUC6-12}

\item{partialConc}{Point estimated concentration Ex:1,4 for concentration after 1 and 4 h)}

\item{plots}{plog, plin, or plot}
}
\description{
nca.cal()
}
\keyword{nca.cal}
