% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nm_ph}
\alias{nm_ph}
\title{Format NONMEM output for phx_typical}
\usage{
nm_ph(ext = "run16.ext", phi = "run16.phi", eta.names = "phi")
}
\arguments{
\item{ext}{theta sheet}

\item{phi}{required for shrinkage calculation}
}
\description{
Format NONMEM output for phx_typical
}
\keyword{nm_ph}
