% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{phx_typical}
\alias{phx_typical}
\title{Summary of typical values from Phoenix}
\usage{
phx_typical(
  theta = th,
  omega = om,
  omega_sd = omsd,
  sd = NULL,
  estimate = "Estimate",
  lab = c("tvKa;;Ka (1/h);;x+y;;c(0,0);;nKa",
    "dKadSTR100;;If dose=100;;exp(x)+y;;c(0,0);;", "tvCl;;CL/F (L/h);;x+y;;c(0,0);;nCl",
    "dCldMULT2;;CL, if multiple dose;;exp(x)*y;;tvCl;;",
    "dCldECOG1;;CL, if ECOG=1;;exp(x)*y;;tvCl;;",
    "stdev0;;Proportional Error (\%);;x*100+y;;c(0,0);;")
)
}
\arguments{
\item{theta}{theta sheet}

\item{omega}{omega data matrix format}

\item{omega_sd}{sd for omega matrix format}

\item{sd}{set to null if no covariance steps or specify variable name for SD theta}

\item{estimate}{specify variable name for estimate theta}

\item{lab}{table format to be outputted (see example)}
}
\description{
Summary of typical values from Phoenix
}
\examples{
Output table to be edited in lab: Require 1) original Param name ;; 2) explicit name;;3)transformation equation (up to 2 variables x and y)
variables);;4) y values if needed or use original variable names (ex:CL) or set to c(0,0) if not used ;; 5)names corresponding ETA (ex: nCL)
Note: each variable in the lab is separated by ;;. If no transformation of x needed, 3) x+y and 4) c(0,0). Note that the error sd will be propagated 
according to the equation.
}
\keyword{phx_typical}
