% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{phx_typical}
\alias{phx_typical}
\title{Summary of typical values from Phoenix}
\usage{
phx_typical(
  theta = th,
  omega = om,
  omega_sd = NULL,
  sd = NULL,
  estimate = "Estimate",
  lab = c("tvKa;;Ka (1/h);;x+y;;c(0,0);;nKa;;sqrt(exp(x)-1)*100",
    "dKadSTR100;;If dose=100;;exp(x)+y;;c(0,0);;",
    "tvCl;;CL/F (L/h);;x+y;;c(0,0);;nCl;;sqrt(exp(x)-1)*100",
    "dCldMULT2;;CL, if multiple dose;;exp(x)*y;;tvCl;;",
    "tvV;;Vc/F (L);;x+y;;c(0,0);;nV;;sqrt(exp(x)-1)*100",
    "tvCl2;;Q/F (L/h);;x+y;;c(0,0);;nCl2;;sqrt(exp(x)-1)*100",
    "tvV2;;Vp/F (L/h);;x+y;;c(0,0);;nV2;;sqrt(exp(x)-1)*100",
    "stdev0;;Proportional Error (\%);;x*100+y;;c(0,0);;")
)
}
\arguments{
\item{theta}{theta sheet}

\item{omega}{omega data matrix format}

\item{omega_sd}{sd for omega matrix format}

\item{sd}{set to null if no covariance steps or specify variable name for SD theta}

\item{estimate}{specify variable name for estimate theta}

\item{lab}{table format to be outputted. Edit the lab list: Require 1) original Param name ;; 2) explicit name;;3)transformation equation (up to 2 variables x and y) variables);;4) specify y c(value,sd) if needed or use original variable names (ex:CL) or set it to c(0,0) if not used ;; 5)names corresponding ETA (ex: nCL);; 6) expression for eta distribution. Note: each variable in the lab is separated by ;;. If no transformation of x needed set 3) to x+y and 4) as c(0,0). Note that the error sd will be derived using error propagation based on the expression.}
}
\description{
Summary of typical values from Phoenix
}
\keyword{phx_typical}
