% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{ppt}
\alias{ppt}
\title{Create Power point}
\usage{
ppt(
  template = "C:/Users/lpheng/Desktop/Templates and Documents/templateforofficer.pptx",
  cover = NULL,
  plain1 = NULL,
  plain2 = NULL
)
}
\arguments{
\item{template}{Required typical template by Leon}

\item{cover}{Cover slide c("Title","Authors","Project No or Date").}

\item{plain1}{Slide with title c("zzz-yyy","Title") and single ggplot figure (C25/C75/C100), image (P25/P75/P100) or table (TAB). Sizes are 25%,75% and 100%.}

\item{plain2}{Slide with title c("zzz-yyy","Title") and combination of figure (Text or paragraph. For heading 1, 2 or 3 text c("text,1, 2 or 3)}
}
\description{
Create Power point
}
\examples{
pres<-NULL
pres<-ppt()
pres<-ppt(cover=c("zzz","yyy","rrr")))
pres<-ppt(plain1=list())
pres<-ppt(plain2=list())
pres<-ppt(plain2=list())
print(pres,"test..pptx")
}
\keyword{ppt}
