% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{pptdoc}
\alias{pptdoc}
\title{Create Power point}
\usage{
pptdoc(
  template = "C:/Users/lpheng/Desktop/Templates and Documents/templateforofficer.pptx",
  cover = NULL,
  plain1 = NULL,
  plain2 = NULL
)
}
\arguments{
\item{template}{Required typical template by Leon}

\item{cover}{Cover slide c("Title","Authors","Project No or Date").}

\item{plain1}{Slide with title c("zzz-yyy","Title") and single ggplot figure (C25/C75/C100), image (P25/P75/P100) or table (TAB). Sizes are 25%,75% and 100%.}

\item{plain2}{Slide with title c("zzz-yyy","Title") and combination of figure (Text or paragraph. For heading 1, 2 or 3 text c("text,1, 2 or 3)}
}
\description{
Create Power point
}
\examples{
pres<-NULL
pres<-pptdoc()
pres<-pptdoc(cover=c("zzz","yyy","rrr")))
pres<-pptdoc(plain1=list())
pres<-pptdoc(plain2=list())
pres<-pptdoc(plain2=list())
print(pres,"test..pptx")
}
\keyword{pptdoc}
