% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{tadRT}
\alias{tadRT}
\title{Derive TAD and RTIME}
\usage{
tadRT(
  data,
  id = "ID",
  dttm = NULL,
  cdate = NULL,
  ctime = NULL,
  evid = "EVID",
  tz = "GMT",
  format = "\%Y-\%m-\%d \%H:\%M"
)
}
\arguments{
\item{data}{data}

\item{id}{subject id}

\item{date}{date variable}

\item{time}{time variable}

\item{EVID}{evid variable}
}
\description{
Derive TAD and RTIME from calendar date and time or dttm
}
\keyword{tadRT}
