% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{wdoc}
\alias{wdoc}
\title{Create word doc}
\usage{
wdoc(
  template = NULL,
  df = NULL,
  tab = NULL,
  txt = NULL,
  img = NULL,
  fig = NULL
)
}
\arguments{
\item{template}{Document template if available, provide full path and template file name}

\item{df}{Data frame will be converted by flextable table}

\item{tab}{Table created by flextable}

\item{txt}{Text or paragraph. For heading 1, 2 or 3 text c("text",1, 2 or 3)}

\item{img}{Plots saved as png. Enter the full path and file name}

\item{fig}{Plot created by ggplot}
}
\description{
Create word doc
}
\examples{
df<-data.frame(x=2,z=4)
ft<-flextable(df)
library(ggplot2)
p<-ggplot(df,aes(x=x,y=z))+
 geom_point()

doc<-NULL
doc<-wdoc()
doc<-wdoc(txt=c("this is a test",1))
doc<-wdoc(df=df)
doc<-wdoc(tab=ft)
doc<-wdoc(fig=p)
print(doc,"test.docx")

}
\keyword{wdoc}
